Carry SPDIF input at 24-bit end-to-end (remove 16-bit truncation)
The SPDIF path truncated twice to 16-bit: the demux dropped the low 8 bits into
an int16 ring, and the resampler's float output was requantized to int16 in the
shared pop_scratch. That capped SPDIF at ~16-bit (~93 dB, undithered) for no
reason — the resampler and DSP graph are float.

Now the path is full-depth, float-staged:
- Demux keeps the sign-extended 24-bit sample; ring is int32 (8 KB, fits the
  existing 0x24012000-0x24014000 AXI hole exactly).
- pop converts the ring to float via 1/8388608 and hands the resampler's native
  float output straight to the DSP — no integer requantization on SPDIF at all.
- Unified the input staging on float: usb_ring_pop_frames and
  spdif_input_pop_frames output float [-1,1], pop_scratch is float, and the DSP
  input stage drops the 1/32768 scaling (and dead INT16_RECIP). USB (16-bit)
  converts via 1/32768.

Gain preserved: 0 dBFS -> ~1.0 exactly as before (verified on hardware: SPDIF
locks at 48 kHz, USB enumerates, audio clean at correct level on both). The only
remaining depth loss on SPDIF is the resampler itself. BSS +0.8 KB (float
pop_scratch); output stage already clips to [-1,1] so resampler overshoot is
handled with float headroom through the graph.

+++ firmware/STM32/Core/Inc/spdif_input.h
@@ -63,12 +63,14 @@ void spdif_input_stop(void);
 uint32_t spdif_input_poll(void);
 
 /* Stage 2: drain up to want_frames stereo frames from the SPDIFRX
- * input ring into dst[2 × want_frames] (interleaved L/R int16, same
- * format as usb_ring_pop_frames). Returns frames actually written;
+ * input ring into dst[2 × want_frames] (interleaved L/R float in [-1, 1],
+ * same format as usb_ring_pop_frames). The SPDIF path carries the full
+ * 24-bit input depth end-to-end — the resampler's float output is handed
+ * over without integer requantization. Returns frames actually written;
  * shortfall is silence-padded so the caller never has to handle a
  * partial fill. Called from the SAI1 DMA half-cplt context — must
  * be lock-free vs the SPDIFRX DMA callbacks that fill the ring. */
-uint32_t spdif_input_pop_frames(int16_t *dst, uint32_t want_frames);
+uint32_t spdif_input_pop_frames(float *dst, uint32_t want_frames);
 
 /* Populate the 16-byte status packet for vendor cmd 0xE2. */
 void spdif_input_get_status(SpdifRxStatusPacket *out);
